[ARCH.MOD.SESSION] Split admission and dispatch modules
