+++ crates/wow-world/src/session_tests.rs
@@ -11,6 +11,15 @@
 
 #![cfg(test)]
 
+#[path = "session/tests/admission.rs"]
+mod admission;
+#[path = "session/tests/dispatch.rs"]
+mod dispatch;
+#[path = "session/tests/routing.rs"]
+mod routing;
+
+use routing::assert_destroyed_party_update_like_cpp;
+
 use super::*;
 use std::sync::atomic::{AtomicU64, AtomicUsize, Ordering as AtomicOrdering};
 use wow_constants::ItemModType;
@@ -8864,165 +8873,6 @@ async fn party_update_command_consumes_receiver_sequence_like_cpp() {
 }
 
 #[tokio::test]
-async fn realm_only_party_commands_never_use_instance_after_connect_to_like_cpp() {
-    let (mut session, _, instance_rx) = make_session();
-    let (realm_tx, realm_rx) = flume::unbounded();
-    session.install_realm_send_channel_for_test(realm_tx);
-    let player_guid = ObjectGuid::create_player(1, 42);
-    let group_registry = Arc::new(GroupRegistry::default());
-    let group = GroupInfo::new(player_guid);
-    let group_guid = group.group_guid;
-    group_registry.insert(group_guid, group);
-    session.group_guid = Some(group_guid);
-    session.set_group_registry(group_registry, Arc::new(PendingInvites::default()));
-    session.set_player_guid(Some(player_guid));
-    session.state = SessionState::LoggedIn;
-
-    let member_full_state = vec![0x59, 0x27, 0xAA];
-    session
-        .session_command_tx()
-        .try_send(SessionCommand::SendPartyUpdateLikeCpp(
-            SendPartyUpdateLikeCppCommand {
-                recipient: player_guid,
-                party_update: wow_packet::packets::party::PartyUpdate {
-                    party_flags: 0,
-                    party_index: wow_network::group_registry::GROUP_CATEGORY_HOME_LIKE_CPP,
-                    party_type: wow_network::group_registry::GROUP_TYPE_NORMAL_LIKE_CPP,
-                    my_index: 0,
-                    party_guid: group_guid,
-                    sequence_num: 999,
-                    leader_guid: player_guid,
-                    leader_faction_group: 0,
-                    player_list: Vec::new(),
-                    loot_settings: None,
-                    difficulty_settings: None,
-                },
-                member_full_state_packets: vec![member_full_state.clone()],
-            },
-        ))
-        .unwrap();
-    session
-        .process_represented_session_commands_like_cpp()
-        .await;
-
-    let update = realm_rx.try_recv().expect("realm PartyUpdate");
-    assert_eq!(
-        u16::from_le_bytes([update[0], update[1]]),
-        ServerOpcodes::PartyUpdate as u16
-    );
-    assert_eq!(realm_rx.try_recv().unwrap(), member_full_state);
-    assert!(instance_rx.try_recv().is_err());
-
-    let invite = vec![0xBD, 0x25, 0xCC];
-    session
-        .session_command_tx()
-        .try_send(SessionCommand::SendRealmPacketLikeCpp(
-            SendRealmPacketLikeCppCommand {
-                recipient: player_guid,
-                packet_bytes: invite.clone(),
-            },
-        ))
-        .unwrap();
-    session
-        .process_represented_session_commands_like_cpp()
-        .await;
-
-    assert_eq!(realm_rx.try_recv().unwrap(), invite);
-    assert!(instance_rx.try_recv().is_err());
-
-    let wrong_recipient = ObjectGuid::create_player(1, 43);
-    session
-        .session_command_tx()
-        .try_send(SessionCommand::SendRealmPacketLikeCpp(
-            SendRealmPacketLikeCppCommand {
-                recipient: wrong_recipient,
-                packet_bytes: invite.clone(),
-            },
-        ))
-        .unwrap();
-    session
-        .process_represented_session_commands_like_cpp()
-        .await;
-    assert!(realm_rx.try_recv().is_err());
-
-    session
-        .session_command_tx()
-        .try_send(SessionCommand::SendPartyUpdateLikeCpp(
-            SendPartyUpdateLikeCppCommand {
-                recipient: wrong_recipient,
-                party_update: wow_packet::packets::party::PartyUpdate {
-                    party_flags: 0,
-                    party_index: wow_network::group_registry::GROUP_CATEGORY_HOME_LIKE_CPP,
-                    party_type: wow_network::group_registry::GROUP_TYPE_NORMAL_LIKE_CPP,
-                    my_index: 0,
-                    party_guid: group_guid,
-                    sequence_num: 0,
-                    leader_guid: player_guid,
-                    leader_faction_group: 0,
-                    player_list: Vec::new(),
-                    loot_settings: None,
-                    difficulty_settings: None,
-                },
-                member_full_state_packets: Vec::new(),
-            },
-        ))
-        .unwrap();
-    session
-        .process_represented_session_commands_like_cpp()
-        .await;
-    assert!(realm_rx.try_recv().is_err());
-
-    session.state = SessionState::Authed;
-    session
-        .session_command_tx()
-        .try_send(SessionCommand::SendRealmPacketLikeCpp(
-            SendRealmPacketLikeCppCommand {
-                recipient: player_guid,
-                packet_bytes: invite,
-            },
-        ))
-        .unwrap();
-    session
-        .process_represented_session_commands_like_cpp()
-        .await;
-    assert!(realm_rx.try_recv().is_err());
-    assert!(instance_rx.try_recv().is_err());
-}
-
-/// Parses the destroyed `PartyUpdate` that C++
-/// `Group::SendUpdateDestroyGroupToPlayer` (`Group.cpp:917-926`) sends so
-/// the removed member's client tears down its party frames.
-fn assert_destroyed_party_update_like_cpp(bytes: &[u8], group_guid: u64) {
-    let mut packet = WorldPacket::from_bytes(bytes);
-    assert_eq!(
-        packet.read_uint16().expect("opcode"),
-        ServerOpcodes::PartyUpdate as u16
-    );
-    assert_eq!(
-        packet.read_uint16().expect("party flags"),
-        wow_network::group_registry::GROUP_FLAG_DESTROYED_LIKE_CPP
-    );
-    assert_eq!(
-        packet.read_uint8().expect("party index"),
-        wow_network::group_registry::GROUP_CATEGORY_HOME_LIKE_CPP
-    );
-    assert_eq!(
-        packet.read_uint8().expect("party type"),
-        wow_network::group_registry::GROUP_TYPE_NONE_LIKE_CPP
-    );
-    assert_eq!(packet.read_int32().expect("my index"), -1);
-    assert_eq!(
-        packet.read_packed_guid().expect("party guid"),
-        ObjectGuid::create_group(group_guid)
-    );
-    let _sequence_num = packet.read_int32().expect("sequence num");
-    assert_eq!(
-        packet.read_packed_guid().expect("leader guid"),
-        ObjectGuid::EMPTY
-    );
-}
-
-#[tokio::test]
 async fn group_removal_command_clears_remote_party_type_like_cpp() {
     let (mut session, _, instance_rx) = make_session();
     let (realm_tx, realm_rx) = flume::bounded(8);
@@ -11022,38 +10872,6 @@ async fn send_if_visible_command_rejected_on_wrong_map_id_like_cpp() {
     assert!(
         send_rx.try_recv().is_err(),
         "must not send when map_id does not match"
-    );
-}
-
-/// (4) Packet is NOT sent when `instance_id` in command does not match session instance.
-/// Slice 4A.1b requirement — instance separation, no cross-instance delivery.
-#[tokio::test]
-async fn send_if_visible_command_rejected_on_wrong_instance_id_like_cpp() {
-    let (mut session, _, send_rx) = make_session();
-    let source_guid = ObjectGuid::create_world_object(HighGuid::Creature, 0, 1, 571, 0, 777, 1004);
-    session.state = SessionState::LoggedIn;
-    session.set_player_map_position_like_cpp(571, Position::ZERO);
-    // session has no canonical map manager → instance_id fallback is 0
-    session.client_visible_guids_like_cpp.insert(source_guid);
-
-    session
-        .session_command_tx()
-        .try_send(SessionCommand::SendIfVisibleLikeCpp(
-            SendIfVisibleLikeCppCommand {
-                queued_at: Instant::now(),
-                source_guid,
-                map_id: 571,
-                instance_id: 99, // different instance
-                packet_bytes: vec![0x88],
-            },
-        ))
-        .expect("command queued");
-    session
-        .process_represented_session_commands_like_cpp()
-        .await;
-    assert!(
-        send_rx.try_recv().is_err(),
-        "must not send when instance_id does not match"
     );
 }
 
@@ -22104,211 +21922,6 @@ fn time_sync_response_sets_initial_clock_delta_like_cpp() {
         session.time_sync_clock_delta >= 1,
         "expected initial fallback delta from first sample"
     );
-}
-
-#[tokio::test]
-async fn time_sync_response_alias_opcodes_dispatch_to_same_handler_like_cpp() {
-    for opcode in [
-        ClientOpcodes::TimeSyncResponseDropped,
-        ClientOpcodes::TimeSyncResponseFailed,
-    ] {
-        let (mut session, _pkt_tx, _send_rx) = make_session();
-        session.set_state(SessionState::LoggedIn);
-        let _ = WorldSession::game_time_ms_like_cpp();
-        std::thread::sleep(std::time::Duration::from_millis(2));
-        let sent_time = WorldSession::game_time_ms_like_cpp();
-        session.time_sync_pending_requests.insert(11, sent_time);
-
-        let mut bytes = Vec::new();
-        bytes.extend_from_slice(&(opcode as u16).to_le_bytes());
-        bytes.extend_from_slice(&11u32.to_le_bytes());
-        bytes.extend_from_slice(&sent_time.saturating_sub(1).to_le_bytes());
-
-        session
-            .dispatch_packet(WorldPacket::from_bytes(&bytes))
-            .await;
-
-        assert!(
-            session.time_sync_pending_requests.is_empty(),
-            "{opcode:?} must consume the pending request like C++ HandleTimeSyncResponse"
-        );
-        assert_eq!(
-            session.time_sync_clock_delta_queue.len(),
-            1,
-            "{opcode:?} must record one clock-delta sample through the shared handler"
-        );
-    }
-}
-
-#[tokio::test]
-async fn party_uninvite_wire_dispatch_parses_packed_guid_and_honors_logged_in_gate_like_cpp() {
-    fn packet(target: ObjectGuid) -> WorldPacket {
-        let mut packet = WorldPacket::new_empty();
-        packet.write_uint16(ClientOpcodes::PartyUninvite as u16);
-        packet.write_bit(false);
-        packet.write_bits(3, 8);
-        packet.write_packed_guid(&target);
-        packet.write_string("bye");
-        packet.flush_bits();
-        packet.reset_read();
-        packet
-    }
-
-    let player = ObjectGuid::create_player(1, 101);
-    let target = ObjectGuid::create_player(1, 202);
-    let (mut logged_in, _pkt_tx, send_rx) = make_session();
-    logged_in.set_state(SessionState::LoggedIn);
-    logged_in.set_player_guid(Some(player));
-
-    logged_in.dispatch_packet(packet(target)).await;
-
-    let result = send_rx.try_recv().expect("PartyCommandResult");
-    assert_eq!(
-        u16::from_le_bytes([result[0], result[1]]),
-        ServerOpcodes::PartyCommandResult as u16
-    );
-    assert!(send_rx.try_recv().is_err());
-
-    let (mut authed, _pkt_tx, send_rx) = make_session();
-    authed.set_player_guid(Some(player));
-    authed.dispatch_packet(packet(target)).await;
-    assert!(
-        send_rx.try_recv().is_err(),
-        "LoggedIn metadata must reject PartyUninvite while the session is Authed"
-    );
-}
-
-#[tokio::test]
-async fn move_set_vehicle_rec_id_ack_wire_dispatch_reaches_handler_only_when_logged_in() {
-    fn packet(mover: ObjectGuid) -> WorldPacket {
-        let mut packet = WorldPacket::new_empty();
-        packet.write_uint16(ClientOpcodes::MoveSetVehicleRecIdAck as u16);
-        wow_packet::packets::movement::MovementInfo {
-            guid: mover,
-            time: 12_345,
-            position: Position::new(1.25, -2.5, 3.75, 0.5),
-            ..Default::default()
-        }
-        .write(&mut packet);
-        packet.write_int32(-77);
-        packet.write_int32(9_001);
-        packet.reset_read();
-        packet
-    }
-
-    use crate::handlers::movement::take_move_set_vehicle_rec_id_ack_handler_calls_for_test;
-
-    let mover = ObjectGuid::create_player(1, 101);
-    assert_eq!(take_move_set_vehicle_rec_id_ack_handler_calls_for_test(), 0);
-
-    let (mut logged_in, _pkt_tx, send_rx) = make_session();
-    logged_in.set_state(SessionState::LoggedIn);
-    logged_in.set_player_guid(Some(mover));
-    logged_in.dispatch_packet(packet(mover)).await;
-
-    assert_eq!(take_move_set_vehicle_rec_id_ack_handler_calls_for_test(), 1);
-    assert!(
-        send_rx.try_recv().is_err(),
-        "C++ HandleMoveSetVehicleRecAck sends no response"
-    );
-
-    let (mut authed, _pkt_tx, send_rx) = make_session();
-    authed.set_player_guid(Some(mover));
-    authed.dispatch_packet(packet(mover)).await;
-
-    assert_eq!(
-        take_move_set_vehicle_rec_id_ack_handler_calls_for_test(),
-        0,
-        "LoggedIn metadata must reject the ACK while the session is Authed"
-    );
-    assert!(send_rx.try_recv().is_err());
-}
-
-#[test]
-fn reconciled_handler_registrations_match_cpp_metadata_and_rust_targets() {
-    let (session, _, _) = make_session();
-    let expected = [
-        (
-            ClientOpcodes::TrainerBuySpell,
-            SessionStatus::LoggedIn,
-            PacketProcessing::Inplace,
-            "handle_trainer_buy_spell",
-        ),
-        (
-            ClientOpcodes::PartyUninvite,
-            SessionStatus::LoggedIn,
-            PacketProcessing::ThreadUnsafe,
-            "handle_party_uninvite",
-        ),
-        (
-            ClientOpcodes::MoveSetVehicleRecIdAck,
-            SessionStatus::LoggedIn,
-            PacketProcessing::ThreadSafe,
-            "handle_move_set_vehicle_rec_id_ack",
-        ),
-    ];
-
-    for (opcode, status, processing, handler_name) in expected {
-        let entry = session
-            .dispatch_table
-            .get(&opcode)
-            .unwrap_or_else(|| panic!("missing linked registration for {opcode:?}"));
-        assert_eq!(entry.status, status, "{opcode:?} status");
-        assert_eq!(entry.processing, processing, "{opcode:?} processing");
-        assert_eq!(entry.handler_name, handler_name, "{opcode:?} target");
-    }
-}
-
-#[tokio::test]
-async fn dispatch_routes_send_text_emote_to_handler_like_cpp() {
-    let (mut session, _pkt_tx, send_rx) = make_session();
-    let player_guid = ObjectGuid::create_player(1, 101);
-    session.set_state(SessionState::LoggedIn);
-    session.set_player_guid(Some(player_guid));
-    session.player_name = Some("Emoter".to_string());
-    session.set_emotes_text_store_like_cpp(Arc::new(wow_data::EmotesTextStore::from_entries([
-        wow_data::EmotesTextEntry {
-            id: 101,
-            name: "wave".to_string(),
-            emote_id: 3,
-        },
-    ])));
-
-    let mut packet = WorldPacket::new_empty();
-    packet.write_uint16(ClientOpcodes::SendTextEmote as u16);
-    packet.write_packed_guid(&ObjectGuid::EMPTY);
-    packet.write_int32(101);
-    packet.write_int32(-1);
-    packet.write_uint32(0);
-    packet.write_int32(0);
-    let bytes = packet.data().to_vec();
-
-    session
-        .dispatch_packet(WorldPacket::from_bytes(&bytes))
-        .await;
-
-    let mut anim = WorldPacket::from_bytes(&send_rx.try_recv().expect("anim emote"));
-    assert_eq!(
-        anim.read_uint16().expect("anim opcode"),
-        ServerOpcodes::Emote as u16
-    );
-    assert_eq!(anim.read_packed_guid().expect("anim source"), player_guid);
-    assert_eq!(anim.read_int32().expect("anim emote"), 3);
-
-    let mut text = WorldPacket::from_bytes(&send_rx.try_recv().expect("text emote"));
-    assert_eq!(
-        text.read_uint16().expect("text opcode"),
-        ServerOpcodes::TextEmote as u16
-    );
-    assert_eq!(text.read_packed_guid().expect("text source"), player_guid);
-    let _account_guid = text.read_packed_guid().expect("source account");
-    assert_eq!(text.read_int32().expect("text emote id"), 101);
-    assert_eq!(text.read_int32().expect("text sound index"), -1);
-    assert_eq!(
-        text.read_packed_guid().expect("text target"),
-        ObjectGuid::EMPTY
-    );
-    assert!(send_rx.try_recv().is_err());
 }
 
 #[test]
@@ -60827,586 +60440,6 @@ fn player_currency_remove_and_save_state_match_cpp() {
 }
 
 #[test]
-fn update_empty_queue() {
-    let (mut session, _, _) = make_session();
-    let processed = session.update(100);
-    assert_eq!(processed, 0);
-}
-
-#[test]
-fn update_processes_packets() {
-    let (mut session, pkt_tx, _) = make_session();
-
-    // Send some packets (they'll be logged as "no handler" but won't crash)
-    for _ in 0..5 {
-        let pkt = WorldPacket::from_bytes(&[0x00, 0x00]); // opcode 0
-        pkt_tx.send(pkt).unwrap();
-    }
-
-    let processed = session.update(100);
-    assert_eq!(processed, 5);
-    assert_eq!(session.pending_packets.len(), 5);
-}
-
-#[test]
-fn update_disconnects_when_socket_timeout_deadline_expired_like_cpp() {
-    let (mut session, _, _) = make_session();
-    session.set_socket_timeouts_like_cpp(SocketTimeoutsLikeCpp {
-        unauthenticated_secs: 60,
-        active_secs: 30,
-    });
-    session.socket_timeout_deadline_like_cpp = Instant::now() - Duration::from_secs(1);
-
-    assert_eq!(session.update(100), 0);
-    assert!(session.is_disconnecting());
-}
-
-#[test]
-fn timed_logout_preserves_player_until_disconnect_save_like_cpp() {
-    let (mut session, _, send_rx) = make_session();
-    let guid = ObjectGuid::create_player(1, 77);
-    session.set_player_guid(Some(guid));
-    session.set_state(SessionState::LoggedIn);
-    session.logout_time = Some(Instant::now() - Duration::from_secs(1));
-
-    session.update(100);
-
-    let packet = send_rx.try_recv().expect("LogoutComplete packet");
-    let mut packet = WorldPacket::from_bytes(&packet);
-    assert_eq!(
-        packet.read_uint16().unwrap(),
-        wow_constants::ServerOpcodes::LogoutComplete as u16
-    );
-    assert_eq!(session.player_guid(), Some(guid));
-    assert!(session.is_disconnecting());
-}
-
-#[test]
-fn update_resets_socket_timeout_on_regular_packet_like_cpp() {
-    let (mut session, pkt_tx, _) = make_session();
-    session.set_socket_timeouts_like_cpp(SocketTimeoutsLikeCpp {
-        unauthenticated_secs: 60,
-        active_secs: 30,
-    });
-    session.socket_timeout_deadline_like_cpp = Instant::now() - Duration::from_secs(1);
-    pkt_tx
-        .send(WorldPacket::from_bytes(&[0x00, 0x00]))
-        .expect("packet queued");
-
-    assert_eq!(session.update(100), 1);
-    assert!(!session.is_disconnecting());
-}
-
-#[test]
-fn keep_alive_only_resets_socket_timeout_for_logged_in_session_like_cpp() {
-    let keep_alive_opcode = (ClientOpcodes::KeepAlive as u16).to_le_bytes();
-
-    let (mut authed_session, authed_tx, _) = make_session();
-    authed_session.set_socket_timeouts_like_cpp(SocketTimeoutsLikeCpp {
-        unauthenticated_secs: 60,
-        active_secs: 30,
-    });
-    authed_session.socket_timeout_deadline_like_cpp = Instant::now() - Duration::from_secs(1);
-    authed_tx
-        .send(WorldPacket::from_bytes(&keep_alive_opcode))
-        .expect("keepalive queued");
-
-    assert_eq!(authed_session.update(100), 1);
-    assert!(authed_session.is_disconnecting());
-
-    let (mut logged_in_session, logged_in_tx, _) = make_session();
-    logged_in_session.set_state(SessionState::LoggedIn);
-    logged_in_session.set_socket_timeouts_like_cpp(SocketTimeoutsLikeCpp {
-        unauthenticated_secs: 60,
-        active_secs: 30,
-    });
-    logged_in_session.socket_timeout_deadline_like_cpp = Instant::now() - Duration::from_secs(1);
-    logged_in_tx
-        .send(WorldPacket::from_bytes(&keep_alive_opcode))
-        .expect("keepalive queued");
-
-    assert_eq!(logged_in_session.update(100), 1);
-    assert!(!logged_in_session.is_disconnecting());
-}
-
-#[test]
-fn packet_spoof_policy_kick_blocks_over_limit_opcode_like_cpp() {
-    let hotfix_opcode = (ClientOpcodes::HotfixRequest as u16).to_le_bytes();
-    let (mut session, pkt_tx, _) = make_session();
-    session.set_packet_spoof_config_like_cpp(PacketSpoofConfigLikeCpp {
-        policy: PacketSpoofConfigLikeCpp::POLICY_KICK,
-        ban_mode: PacketSpoofConfigLikeCpp::BAN_ACCOUNT,
-        ban_duration_secs: 86_400,
-    });
-    pkt_tx
-        .send(WorldPacket::from_bytes(&hotfix_opcode))
-        .expect("first packet queued");
-    pkt_tx
-        .send(WorldPacket::from_bytes(&hotfix_opcode))
-        .expect("second packet queued");
-
-    assert_eq!(session.update(100), 1);
-    assert_eq!(session.pending_packets.len(), 1);
-    assert!(session.is_disconnecting());
-}
-
-#[test]
-fn packet_spoof_policy_log_keeps_over_limit_opcode_like_cpp() {
-    let hotfix_opcode = (ClientOpcodes::HotfixRequest as u16).to_le_bytes();
-    let (mut session, pkt_tx, _) = make_session();
-    session.set_packet_spoof_config_like_cpp(PacketSpoofConfigLikeCpp {
-        policy: PacketSpoofConfigLikeCpp::POLICY_LOG,
-        ban_mode: PacketSpoofConfigLikeCpp::BAN_ACCOUNT,
-        ban_duration_secs: 86_400,
-    });
-    pkt_tx
-        .send(WorldPacket::from_bytes(&hotfix_opcode))
-        .expect("first packet queued");
-    pkt_tx
-        .send(WorldPacket::from_bytes(&hotfix_opcode))
-        .expect("second packet queued");
-
-    assert_eq!(session.update(100), 2);
-    assert_eq!(session.pending_packets.len(), 2);
-    assert!(!session.is_disconnecting());
-}
-
-#[test]
-fn packet_spoof_policy_ban_stages_account_ban_like_cpp() {
-    let hotfix_opcode = (ClientOpcodes::HotfixRequest as u16).to_le_bytes();
-    let (mut session, pkt_tx, _) = make_session();
-    session.set_packet_spoof_config_like_cpp(PacketSpoofConfigLikeCpp {
-        policy: PacketSpoofConfigLikeCpp::POLICY_BAN,
-        ban_mode: PacketSpoofConfigLikeCpp::BAN_ACCOUNT,
-        ban_duration_secs: 3_600,
-    });
-    pkt_tx
-        .send(WorldPacket::from_bytes(&hotfix_opcode))
-        .expect("first packet queued");
-    pkt_tx
-        .send(WorldPacket::from_bytes(&hotfix_opcode))
-        .expect("second packet queued");
-
-    assert_eq!(session.update(100), 1);
-    assert!(session.is_disconnecting());
-    assert_eq!(
-        session.pending_packet_spoof_ban_like_cpp,
-        Some(PacketSpoofPendingBanLikeCpp {
-            target: PacketSpoofPendingBanTargetLikeCpp::Account { account_id: 1 },
-            duration_secs: 3_600,
-        })
-    );
-}
-
-#[test]
-fn packet_spoof_policy_ban_stages_ip_ban_from_remote_address_like_cpp() {
-    let hotfix_opcode = (ClientOpcodes::HotfixRequest as u16).to_le_bytes();
-    let (mut session, pkt_tx, _) = make_session();
-    session.set_remote_address_like_cpp(Some("203.0.113.77".to_string()));
-    session.set_packet_spoof_config_like_cpp(PacketSpoofConfigLikeCpp {
-        policy: PacketSpoofConfigLikeCpp::POLICY_BAN,
-        ban_mode: PacketSpoofConfigLikeCpp::BAN_IP,
-        ban_duration_secs: 7_200,
-    });
-    pkt_tx
-        .send(WorldPacket::from_bytes(&hotfix_opcode))
-        .expect("first packet queued");
-    pkt_tx
-        .send(WorldPacket::from_bytes(&hotfix_opcode))
-        .expect("second packet queued");
-
-    assert_eq!(session.update(100), 1);
-    assert!(session.is_disconnecting());
-    assert_eq!(
-        session.pending_packet_spoof_ban_like_cpp,
-        Some(PacketSpoofPendingBanLikeCpp {
-            target: PacketSpoofPendingBanTargetLikeCpp::Ip {
-                address: "203.0.113.77".to_string(),
-            },
-            duration_secs: 7_200,
-        })
-    );
-}
-
-#[tokio::test]
-async fn kick_like_cpp_session_command_disconnects_session() {
-    let (mut session, _, _) = make_session();
-    session
-        .session_command_tx()
-        .try_send(SessionCommand::KickLikeCpp(KickLikeCppCommand {
-            reason: "World::BanAccount Banning account".to_string(),
-        }))
-        .expect("kick command queued");
-
-    session
-        .process_represented_session_commands_like_cpp()
-        .await;
-
-    assert!(session.is_disconnecting());
-}
-
-#[tokio::test]
-async fn shutdown_flush_command_observes_prior_kick_like_cpp() {
-    let (mut session, _, _) = make_session();
-    let (response_tx, response_rx) = flume::bounded(1);
-    let command_tx = session.session_command_tx();
-
-    command_tx
-        .try_send(SessionCommand::KickLikeCpp(KickLikeCppCommand {
-            reason: "World::KickAll".to_string(),
-        }))
-        .expect("kick command queued");
-    command_tx
-        .try_send(SessionCommand::WorldSessionShutdownFlushLikeCpp(
-            WorldSessionShutdownFlushLikeCppCommand {
-                diff_ms: 1,
-                response_tx,
-            },
-        ))
-        .expect("flush command queued");
-
-    session
-        .process_represented_session_commands_like_cpp()
-        .await;
-
-    let result = response_rx.try_recv().expect("flush ack returned");
-    assert_eq!(result.diff_ms, 1);
-    assert!(result.disconnecting);
-    assert!(session.is_disconnecting());
-}
-
-#[test]
-fn packet_spoof_ban_eviction_queues_kick_for_affected_accounts_like_cpp() {
-    let (mut session, _, _) = make_session();
-    let registry = Arc::new(PlayerRegistry::default());
-    let (send_tx_a, _send_rx_a) = flume::bounded(1);
-    let (send_tx_b, _send_rx_b) = flume::bounded(1);
-    let (send_tx_c, _send_rx_c) = flume::bounded(1);
-    let (command_tx_a, command_rx_a) = flume::bounded(1);
-    let (command_tx_b, command_rx_b) = flume::bounded(1);
-    let (command_tx_c, command_rx_c) = flume::bounded(1);
-
-    let guid_a = ObjectGuid::create_player(1, 11);
-    let guid_b = ObjectGuid::create_player(1, 12);
-    let guid_c = ObjectGuid::create_player(1, 13);
-    let mut info_a = broadcast_info(guid_a, send_tx_a);
-    info_a.account_id = 7;
-    info_a.command_tx = command_tx_a;
-    let mut info_b = broadcast_info(guid_b, send_tx_b);
-    info_b.account_id = 9;
-    info_b.command_tx = command_tx_b;
-    let mut info_c = broadcast_info(guid_c, send_tx_c);
-    info_c.account_id = 7;
-    info_c.command_tx = command_tx_c;
-
-    registry.insert(guid_a, info_a);
-    registry.insert(guid_b, info_b);
-    registry.insert(guid_c, info_c);
-    session.set_player_registry(registry);
-
-    assert_eq!(
-        session.kick_packet_spoof_affected_sessions_like_cpp(&[7]),
-        2
-    );
-    assert!(matches!(
-        command_rx_a.try_recv(),
-        Ok(SessionCommand::KickLikeCpp(_))
-    ));
-    assert!(matches!(
-        command_rx_b.try_recv(),
-        Err(flume::TryRecvError::Empty)
-    ));
-    assert!(matches!(
-        command_rx_c.try_recv(),
-        Ok(SessionCommand::KickLikeCpp(_))
-    ));
-}
-
-#[test]
-fn packet_spoof_cpp_opcode_limit_table_is_exhaustive_like_cpp() {
-    for opcode in [
-        ClientOpcodes::PlayerLogin,
-        ClientOpcodes::QueryPlayerNames,
-        ClientOpcodes::QueryPetName,
-        ClientOpcodes::QueryNpcText,
-        ClientOpcodes::AttackStop,
-        ClientOpcodes::QueryTime,
-        ClientOpcodes::QueryCorpseTransport,
-        ClientOpcodes::MoveTimeSkipped,
-        ClientOpcodes::QueryNextMailTime,
-        ClientOpcodes::SetSheathed,
-        ClientOpcodes::UpdateRaidTarget,
-        ClientOpcodes::LogoutRequest,
-        ClientOpcodes::PetRename,
-        ClientOpcodes::QuestGiverRequestReward,
-        ClientOpcodes::CompleteCinematic,
-        ClientOpcodes::NextCinematicCamera,
-        ClientOpcodes::OpeningCinematic,
-        ClientOpcodes::BankerActivate,
-        ClientOpcodes::BuyBankSlot,
-        ClientOpcodes::OptOutOfLoot,
-        ClientOpcodes::CalendarComplain,
-        ClientOpcodes::QueryQuestInfo,
-        ClientOpcodes::QueryGameObject,
-        ClientOpcodes::QueryCreature,
-        ClientOpcodes::QuestGiverStatusQuery,
-        ClientOpcodes::QueryGuildInfo,
-        ClientOpcodes::TaxiNodeStatusQuery,
-        ClientOpcodes::TaxiQueryAvailableNodes,
-        ClientOpcodes::QuestGiverQueryQuest,
-        ClientOpcodes::QueryPageText,
-        ClientOpcodes::GuildBankTextQuery,
-        ClientOpcodes::QueryCorpseLocationFromClient,
-        ClientOpcodes::MoveSetFacing,
-        ClientOpcodes::MoveSetFacingHeartbeat,
-        ClientOpcodes::MoveSetPitch,
-        ClientOpcodes::RequestPartyMemberStats,
-        ClientOpcodes::QuestGiverCompleteQuest,
-        ClientOpcodes::SetActionButton,
-        ClientOpcodes::SetActionBarToggles,
-        ClientOpcodes::ResetInstances,
-        ClientOpcodes::HearthAndResurrect,
-        ClientOpcodes::TogglePvp,
-        ClientOpcodes::SetPvp,
-        ClientOpcodes::PetAbandon,
-        ClientOpcodes::ActivateTaxi,
-        ClientOpcodes::SelfRes,
-        ClientOpcodes::UnlearnSkill,
-        ClientOpcodes::SaveEquipmentSet,
-        ClientOpcodes::AssignEquipmentSetSpec,
-        ClientOpcodes::DeleteEquipmentSet,
-        ClientOpcodes::RepopRequest,
-        ClientOpcodes::PartyInvite,
-        ClientOpcodes::PartyInviteResponse,
-        ClientOpcodes::PartyUninvite,
-        ClientOpcodes::LeaveGroup,
-        ClientOpcodes::AcceptWargameInvite,
-        ClientOpcodes::BattlemasterJoinArena,
-        ClientOpcodes::BattlemasterHello,
-        ClientOpcodes::BattlefieldList,
-        ClientOpcodes::BattlefieldPort,
-        ClientOpcodes::BattlefieldLeave,
-        ClientOpcodes::BattlemasterJoin,
-        ClientOpcodes::GuildBankLogQuery,
-        ClientOpcodes::LogoutCancel,
-        ClientOpcodes::AlterAppearance,
-        ClientOpcodes::SetPlayerDeclinedNames,
-        ClientOpcodes::QuestConfirmAccept,
-        ClientOpcodes::GuildEventLogQuery,
-        ClientOpcodes::QuestGiverStatusMultipleQuery,
-        ClientOpcodes::InitiateTrade,
-        ClientOpcodes::ChatAddonMessage,
-        ClientOpcodes::ChatAddonMessageWhisper,
-        ClientOpcodes::ChatMessageAfk,
-        ClientOpcodes::ChatMessageChannel,
-        ClientOpcodes::ChatMessageDnd,
-        ClientOpcodes::ChatMessageEmote,
-        ClientOpcodes::ChatMessageGuild,
-        ClientOpcodes::ChatMessageOfficer,
-        ClientOpcodes::ChatMessageParty,
-        ClientOpcodes::ChatMessageRaid,
-        ClientOpcodes::ChatMessageRaidWarning,
-        ClientOpcodes::ChatMessageSay,
-        ClientOpcodes::ChatMessageWhisper,
-        ClientOpcodes::ChatMessageYell,
-        ClientOpcodes::UpdateAadcStatus,
-        ClientOpcodes::Inspect,
-        ClientOpcodes::AreaSpiritHealerQuery,
-        ClientOpcodes::StandStateChange,
-        ClientOpcodes::RandomRoll,
-        ClientOpcodes::TimeSyncResponse,
-        ClientOpcodes::TimeSyncResponseDropped,
-        ClientOpcodes::TimeSyncResponseFailed,
-        ClientOpcodes::MoveForceRunSpeedChangeAck,
-        ClientOpcodes::MoveForceSwimSpeedChangeAck,
-        ClientOpcodes::MoveForceSwimBackSpeedChangeAck,
-        ClientOpcodes::MoveForceRunBackSpeedChangeAck,
-        ClientOpcodes::MoveForceFlightSpeedChangeAck,
-        ClientOpcodes::MoveForceFlightBackSpeedChangeAck,
-        ClientOpcodes::MoveForceWalkSpeedChangeAck,
-        ClientOpcodes::MoveForceTurnRateChangeAck,
-        ClientOpcodes::MoveForcePitchRateChangeAck,
-    ] {
-        assert_eq!(
-            WorldSession::packet_spoof_max_packet_counter_allowed_like_cpp(opcode),
-            0,
-            "{opcode:?}"
-        );
-    }
-
-    for opcode in [
-        ClientOpcodes::QuestGiverAcceptQuest,
-        ClientOpcodes::QuestLogRemoveQuest,
-        ClientOpcodes::QuestGiverChooseReward,
-        ClientOpcodes::SendContactList,
-        ClientOpcodes::AutobankItem,
-        ClientOpcodes::AutostoreBankItem,
-        ClientOpcodes::Who,
-        ClientOpcodes::RideVehicleInteract,
-        ClientOpcodes::MoveHeartbeat,
-    ] {
-        assert_eq!(
-            WorldSession::packet_spoof_max_packet_counter_allowed_like_cpp(opcode),
-            200,
-            "{opcode:?}"
-        );
-    }
-
-    for opcode in [
-        ClientOpcodes::GuildSetMemberNote,
-        ClientOpcodes::SetContactNotes,
-        ClientOpcodes::CalendarGet,
-        ClientOpcodes::GuildBankQueryTab,
-        ClientOpcodes::QueryInspectAchievements,
-        ClientOpcodes::GameObjReportUse,
-        ClientOpcodes::GameObjUse,
-        ClientOpcodes::DeclinePetition,
-    ] {
-        assert_eq!(
-            WorldSession::packet_spoof_max_packet_counter_allowed_like_cpp(opcode),
-            50,
-            "{opcode:?}"
-        );
-    }
-
-    assert_eq!(
-        WorldSession::packet_spoof_max_packet_counter_allowed_like_cpp(
-            ClientOpcodes::QuestPoiQuery
-        ),
-        crate::handlers::quest::MAX_QUEST_LOG_SIZE_LIKE_CPP as u32
-    );
-
-    for opcode in [ClientOpcodes::SpellClick, ClientOpcodes::MoveDismissVehicle] {
-        assert_eq!(
-            WorldSession::packet_spoof_max_packet_counter_allowed_like_cpp(opcode),
-            20,
-            "{opcode:?}"
-        );
-    }
-
-    for opcode in [
-        ClientOpcodes::SignPetition,
-        ClientOpcodes::TurnInPetition,
-        ClientOpcodes::ChangeSubGroup,
-        ClientOpcodes::QueryPetition,
-        ClientOpcodes::CharCustomize,
-        ClientOpcodes::CharRaceOrFactionChange,
-        ClientOpcodes::CharDelete,
-        ClientOpcodes::DelFriend,
-        ClientOpcodes::AddFriend,
-        ClientOpcodes::CharacterRenameRequest,
-        ClientOpcodes::BugReport,
-        ClientOpcodes::SetPartyLeader,
-        ClientOpcodes::ConvertRaid,
-        ClientOpcodes::SetAssistantLeader,
-        ClientOpcodes::MoveChangeVehicleSeats,
-        ClientOpcodes::PetitionBuy,
-        ClientOpcodes::RequestVehiclePrevSeat,
-        ClientOpcodes::RequestVehicleNextSeat,
-        ClientOpcodes::RequestVehicleSwitchSeat,
-        ClientOpcodes::RequestVehicleExit,
-        ClientOpcodes::EjectPassenger,
-        ClientOpcodes::ItemPurchaseRefund,
-        ClientOpcodes::SocketGems,
-        ClientOpcodes::WrapItem,
-        ClientOpcodes::ReportPvpPlayerAfk,
-    ] {
-        assert_eq!(
-            WorldSession::packet_spoof_max_packet_counter_allowed_like_cpp(opcode),
-            10,
-            "{opcode:?}"
-        );
-    }
-
-    for opcode in [
-        ClientOpcodes::CreateCharacter,
-        ClientOpcodes::EnumCharacters,
-        ClientOpcodes::EnumCharactersDeletedByClient,
-        ClientOpcodes::SubmitUserFeedback,
-        ClientOpcodes::SupportTicketSubmitBug,
-        ClientOpcodes::SupportTicketSubmitComplaint,
-        ClientOpcodes::SupportTicketSubmitSuggestion,
-        ClientOpcodes::CalendarAddEvent,
-        ClientOpcodes::CalendarUpdateEvent,
-        ClientOpcodes::CalendarRemoveEvent,
-        ClientOpcodes::CalendarCopyEvent,
-        ClientOpcodes::CalendarInvite,
-        ClientOpcodes::CalendarEventSignUp,
-        ClientOpcodes::CalendarRsvp,
-        ClientOpcodes::CalendarStatus,
-        ClientOpcodes::CalendarModeratorStatus,
-        ClientOpcodes::CalendarRemoveInvite,
-        ClientOpcodes::SetLootMethod,
-        ClientOpcodes::GuildInviteByName,
-        ClientOpcodes::AcceptGuildInvite,
-        ClientOpcodes::GuildDeclineInvitation,
-        ClientOpcodes::GuildLeave,
-        ClientOpcodes::GuildDelete,
-        ClientOpcodes::GuildSetGuildMaster,
-        ClientOpcodes::GuildUpdateMotdText,
-        ClientOpcodes::GuildSetRankPermissions,
-        ClientOpcodes::GuildAddRank,
-        ClientOpcodes::GuildDeleteRank,
-        ClientOpcodes::GuildUpdateInfoText,
-        ClientOpcodes::GuildBankDepositMoney,
-        ClientOpcodes::GuildBankWithdrawMoney,
-        ClientOpcodes::GuildBankBuyTab,
-        ClientOpcodes::GuildBankUpdateTab,
-        ClientOpcodes::GuildBankSetTabText,
-        ClientOpcodes::SaveGuildEmblem,
-        ClientOpcodes::PetitionRenameGuild,
-        ClientOpcodes::ConfirmRespecWipe,
-        ClientOpcodes::SetDungeonDifficulty,
-        ClientOpcodes::SetRaidDifficulty,
-        ClientOpcodes::SetPartyAssignment,
-        ClientOpcodes::DoReadyCheck,
-    ] {
-        assert_eq!(
-            WorldSession::packet_spoof_max_packet_counter_allowed_like_cpp(opcode),
-            3,
-            "{opcode:?}"
-        );
-    }
-
-    assert_eq!(
-        WorldSession::packet_spoof_max_packet_counter_allowed_like_cpp(
-            ClientOpcodes::GetItemPurchaseData
-        ),
-        PLAYER_SLOT_END as u32
-    );
-    assert_eq!(
-        WorldSession::packet_spoof_max_packet_counter_allowed_like_cpp(
-            ClientOpcodes::HotfixRequest
-        ),
-        1
-    );
-    assert_eq!(
-        WorldSession::packet_spoof_max_packet_counter_allowed_like_cpp(ClientOpcodes::AuthSession),
-        100
-    );
-}
-
-#[test]
-fn packet_spoof_zero_limit_opcode_is_unlimited_like_cpp() {
-    let player_login_opcode = (ClientOpcodes::PlayerLogin as u16).to_le_bytes();
-    let (mut session, pkt_tx, _) = make_session();
-    session.set_packet_spoof_config_like_cpp(PacketSpoofConfigLikeCpp {
-        policy: PacketSpoofConfigLikeCpp::POLICY_KICK,
-        ban_mode: PacketSpoofConfigLikeCpp::BAN_ACCOUNT,
-        ban_duration_secs: 86_400,
-    });
-    for _ in 0..5 {
-        pkt_tx
-            .send(WorldPacket::from_bytes(&player_login_opcode))
-            .expect("packet queued");
-    }
-
-    assert_eq!(session.update(100), 5);
-    assert_eq!(session.pending_packets.len(), 5);
-    assert!(!session.is_disconnecting());
-}
-
-#[test]
 fn kick_sets_disconnecting() {
     let (mut session, _, _) = make_session();
     session.kick("test");
@@ -75072,134 +74105,6 @@ fn realm_query_response_uses_realm_list_names_like_cpp() {
 }
 
 #[test]
-fn dispatch_metadata_matches_cpp_for_registered_active_opcodes() {
-    let (session, _, _) = make_session();
-    let table = &session.dispatch_table;
-
-    fn status_from_cpp(value: &str) -> Option<SessionStatus> {
-        match value {
-            "STATUS_AUTHED" => Some(SessionStatus::Authed),
-            "STATUS_LOGGEDIN" => Some(SessionStatus::LoggedIn),
-            "STATUS_TRANSFER" => Some(SessionStatus::Transfer),
-            "STATUS_LOGGEDIN_OR_RECENTLY_LOGGOUT" => Some(SessionStatus::LoggedInOrRecentlyLogout),
-            "STATUS_NEVER" | "STATUS_UNHANDLED" => None,
-            other => panic!("unknown C++ session status {other}"),
-        }
-    }
-
-    fn processing_from_cpp(value: &str) -> PacketProcessing {
-        match value {
-            "PROCESS_INPLACE" => PacketProcessing::Inplace,
-            "PROCESS_THREADUNSAFE" => PacketProcessing::ThreadUnsafe,
-            "PROCESS_THREADSAFE" => PacketProcessing::ThreadSafe,
-            other => panic!("unknown C++ packet processing {other}"),
-        }
-    }
-
-    let cpp_metadata = include_str!("../../../docs/migration/inventory/cpp-client-handlers.tsv");
-    let mut expected = std::collections::HashMap::new();
-    let mut cpp_never_or_unhandled = std::collections::HashSet::new();
-
-    for line in cpp_metadata.lines().skip(1) {
-        let columns: Vec<_> = line.split('\t').collect();
-        let rust_const = columns[9];
-        if rust_const == "-" {
-            continue;
-        }
-
-        let cpp_status = columns[3];
-        let cpp_processing = columns[4];
-        if let Some(status) = status_from_cpp(cpp_status) {
-            expected.insert(rust_const, (status, processing_from_cpp(cpp_processing)));
-        } else {
-            cpp_never_or_unhandled.insert(rust_const);
-        }
-    }
-
-    let compatibility_exceptions = [
-        "BattlePayGetPurchaseList",
-        "AddBattlenetFriend",
-        "BattlenetChallengeResponse",
-        "ChangeBagSlotFlag",
-        "CloseQuestChoice",
-        "ConnectToFailed",
-        "GetAccountCharacterList",
-        "GetAccountNotifications",
-        "LogStreamingError",
-        "LogoutInstant",
-        "MoveAddImpulseAck",
-        "MoveApplyInertiaAck",
-        "MoveRemoveInertiaAck",
-        "MoveRemoveMovementForces",
-        "MoveSeamlessTransferComplete",
-        "MoveSetAdvFly",
-        "MoveSetAdvFlyingAddImpulseMaxSpeedAck",
-        "MoveSetAdvFlyingAirFrictionAck",
-        "MoveSetAdvFlyingBankingRateAck",
-        "MoveSetAdvFlyingDoubleJumpVelModAck",
-        "MoveSetAdvFlyingGlideStartMinHeightAck",
-        "MoveSetAdvFlyingLaunchSpeedCoefficientAck",
-        "MoveSetAdvFlyingLiftCoefficientAck",
-        "MoveSetAdvFlyingMaxVelAck",
-        "MoveSetAdvFlyingOverMaxDecelerationAck",
-        "MoveSetAdvFlyingPitchingRateDownAck",
-        "MoveSetAdvFlyingPitchingRateUpAck",
-        "MoveSetAdvFlyingSurfaceFrictionAck",
-        "MoveSetAdvFlyingTurnVelocityThresholdAck",
-        "OverrideScreenFlash",
-        "Ping",
-        "QueryQuestItemUsability",
-        "QueryCountdownTimer",
-        "ReportClientVariables",
-        "ReportEnabledAddons",
-        "ReportFrozenWhileLoadingMap",
-        "ReportKeybindingExecutionCounts",
-        "RequestConquestFormulaConstants",
-        "SaveAccountDataExport",
-        "SetInsertItemsLeftToRight",
-        "SetPreferredCemetery",
-        "SpawnTrackingUpdate",
-        "TimeAdjustmentResponse",
-        "UpdateAreaTriggerVisual",
-        "UpdateClientSettings",
-        "UpdateSpellVisual",
-        "UpdateVasPurchaseStates",
-        "UsedFollow",
-    ];
-    let modern_client_non_wotlk_exceptions = [
-        "BattlePetClearFanfare",
-        "BattlePetRequestJournal",
-        "BattlePetRequestJournalLock",
-        "BattlePetSetBattleSlot",
-        "BattlePetSetFlags",
-        "BattlePetSummon",
-        "BattlePetUpdateDisplayNotify",
-        "BattlePetUpdateNotify",
-        "QueryBattlePetName",
-    ];
-
-    for entry in table.values() {
-        let opcode_name = format!("{:?}", entry.opcode);
-        if modern_client_non_wotlk_exceptions.contains(&opcode_name.as_str()) {
-            continue;
-        }
-        if compatibility_exceptions.contains(&opcode_name.as_str()) {
-            assert!(
-                cpp_never_or_unhandled.contains(opcode_name.as_str()),
-                "{opcode_name} is listed as a compatibility exception but C++ metadata is active"
-            );
-            continue;
-        }
-
-        let (status, processing) = expected
-            .get(opcode_name.as_str())
-            .unwrap_or_else(|| panic!("missing C++ metadata row for {opcode_name}"));
-        assert_eq!(entry.status, *status, "{opcode_name} status");
-        assert_eq!(entry.processing, *processing, "{opcode_name} processing");
-    }
-}
-
-#[test]
 fn gameobject_use_mover_guard_matches_cpp_remote_control_branch() {
     let (mut session, _pkt_tx, _send_rx) = make_session();
     let player_guid = ObjectGuid::create_player(1, 99);
@@ -82281,21 +81186,6 @@ async fn two_sessions_sharing_legacy_map_manager_see_same_creature_state() {
         guids2.contains(&creature_guid),
         "session2 must see creature inserted via shared manager"
     );
-}
-
-#[test]
-fn dispatch_table_has_no_duplicate_registered_opcodes() {
-    let mut counts = std::collections::HashMap::new();
-    for entry in inventory::iter::<PacketHandlerEntry> {
-        *counts.entry(entry.opcode).or_insert(0usize) += 1;
-    }
-
-    let duplicates: Vec<_> = counts
-        .into_iter()
-        .filter_map(|(opcode, count)| (count > 1).then_some((opcode, count)))
-        .collect();
-
-    assert!(duplicates.is_empty(), "duplicate handlers: {duplicates:?}");
 }
 
 // ── RuntimeTickOwner / RuntimeOutput tests (#NEXT.RUNTIME.L3.001) ─────────
